- #894 Inclusão de pasta de versão (V1) para os DTO da API de backend
- #1184 Criação de mecanismo para desabilitar respostas XML, se necessário
- #895 Substituição para utilizar decimal ao invés de double e float
- #1188 Inclusão de método para abrir transação no GDA para os métodos da API

Related work items: #1184, #1188, #893, #894, #895

+++ Api/Glass.API.Backend/Helper/Produtos/Estrategias/CalculoTotal/ValidacaoCalculoTotalSemOperacaoStrategy.cs
@@ -3,7 +3,7 @@
 // </copyright>
 
 using GDA;
-using Glass.API.Backend.Models.Produtos.CalculoTotal;
+using Glass.API.Backend.Models.Produtos.V1.CalculoTotal;
 using System.Web.Http;
 
 namespace Glass.API.Backend.Helper.Produtos.Estrategias.CalculoTotal
@@ -26,7 +26,7 @@ namespace Glass.API.Backend.Helper.Produtos.Estrategias.CalculoTotal
         }
 
         /// <inheritdoc/>
-        public double ObterPercentualDescontoQuantidade(string dadosAdicionaisValidacao)
+        public decimal ObterPercentualDescontoQuantidade(string dadosAdicionaisValidacao)
         {
             return 0;
         }
